CRITICAL: Bidirectional analysis reveals and fixes 5 hidden bugs
BIDIRECTIONAL ANALYSIS COMPLETE - All type mismatches traced and verified

CRITICAL BUGS FOUND AND FIXED:
================================

Bug 1: Float literal in clip_gradients call (Line 1965)
- Changed 10.0f → 10.0 to match double parameter

Bug 2: Gradient norm function used float (Lines 2270-2279)
- Return type: float → double
- Local variables: float → double
- Function: prime_sqrtf → prime_sqrt
- IMPACT: Critical - prevents precision loss in gradient monitoring

Bug 3: Softmax computation used float (Lines 508-533)
- max_logit: float → double
- sum_exp: float → double
- x, prob: float → double
- Function: prime_expf → prime_exp
- Literals: 50.0f → 50.0, 1.0f → 1.0, 0.0f → 0.0
- IMPACT: Critical - softmax is core to gradient computation

Bug 4: Layer norm used float (Lines 424-428)
- mean, var, diff: float → double
- IMPACT: Prevents precision loss in normalization

Bug 5: Attention score used float (Line 456)
- score: float → double
- IMPACT: Maintains precision in attention computation

COMPLETE DATA FLOW VERIFICATION:
=================================

Forward Pass:
  Input → Embeddings (double*) → Hidden (double*) →
  Attention (double*) → FF (double*) → LayerNorm (double) →
  Logits (double*)

Backward Pass:
  Logits (double*) → Softmax grad (double) →
  Hidden grad (double*) → Attention grad (double*) →
  Embedding grad (double*) → local_gradients (double*)

Gradient Flow:
  local_gradients (double*) → validate_gradients (double*) →
  clip_gradients (double*) → accumulated_gradients (double*) →
  Optimizer (double*) → Model parameters (double*)

CRYSTALLINE MATH COMPLIANCE:
============================
✅ prime_isnan(double) - was prime_isnanf(float)
✅ prime_isinf(double) - was prime_isinff(float)
✅ prime_sqrt(double) - was prime_sqrtf(float)
✅ prime_exp(double) - was prime_expf(float)

VERIFICATION RESULTS:
=====================
✅ Build: SUCCESS - Zero errors
✅ Type Safety: COMPLETE - All double* consistent
✅ Training: FUNCTIONAL - System starts and runs
✅ NaN/Inf: NONE DETECTED
✅ 12-fold symmetry: VERIFIED (63 threads)
✅ Hierarchical structure: OPERATIONAL

MASTER_PLAN COMPLIANCE:
=======================
✅ OBJECTIVE 3A: Crystalline Math Everywhere
✅ OBJECTIVE 7: Fix All Build Warnings
✅ RULE 4: Complete Bidirectional Analysis

This represents a complete audit of every type change with full
bidirectional tracing through the entire training pipeline.

+++ include/cllm_training_threaded.h
@@ -166,7 +166,7 @@ int threaded_training_get_sphere_stats(ThreadedTrainingSystem* system,
  * @param system Threaded training system
  * @return Total gradient norm
  */
-float threaded_training_get_gradient_norm(ThreadedTrainingSystem* system);
+double threaded_training_get_gradient_norm(ThreadedTrainingSystem* system);
 
 /**
  * Get number of worker spheres
